requirements.txt
uvicorn==0.13.4
pydantic==1.8.2
requests==2.26.0
statsmodels==0.12.1
pandas==1.1.5
fastapi==0.70.1
python_dateutil==2.8.2
